refactor: Update .clean.home alias in bashrc.sh to include ~/.bashdb_hist file
The .clean.home alias in bashrc.sh has been updated to include the removal of the ~/.bashdb_hist file. This file is no longer needed and can be safely deleted along with other unnecessary files.

Refactor the alias to:
.alias .clean.home='rm -rf ~/.viminfo ~/.wget-hsts ~/.lesshst ~/.bashdb_hist ~/.python_history ~/*-ks.cfg ~/.cache/ ~/.pki/ ~/.oracle_jre_usage/ ~/.config/htop/'

This change improves the functionality of the .clean.home alias by ensuring that all unnecessary files are removed from the home directory.